pyproject.toml
[tool.poetry]
name = "ct-django"
version = "0.1.0"
description = "Code and files for the Connecting Threads project."
authors = ["Jason Heppler <jason.heppler@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
django-htmx = "^1.17.3"
django-environ = "^0.11.2"
django-tailwind = "^3.8.0"
django-allauth = "^0.62.1"
psycopg2 = "^2.9.9"
wagtail = "^6.1"
django-taggit = "^5.0.1"
taggit-selectize = "^2.11.0"
python-dotenv = "^1.0.1"
django-modelcluster = "^6.3"
django-import-export = "^4.0.7"
django-admin-interface = "^0.28.8"
geopy = "^2.4.1"
django-tables2 = "^2.7.0"
django-filter = "^24.2"
fontawesomefree = "^6.6.0"
django-storages = "^1.14.4"
boto3 = "^1.35.27"
backoff = "^2.2.1"
ratelimit = "^2.2.1"
requests = "^2.32.3"
django-unfold = "^0.55.2"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
djhtml = "^3.0.6"
pre-commit = "^3.7.0"
django-debug-toolbar = "^4.3.0"
pytest-django = "^4.8.0"
pylint-django = "^2.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
